Give up on non-panning starfield for now.
ParallaxBackground mechanism doesn't honor camera
rotation. :-(

Saw a suggestion of putting the sky as child of a
CanvasItem in the world, which would prevent all
camera transform of the sky, but I couldn't find
the stars at all then, maybe they weren't even rendering.

Anyhow, failed to figure it out.