Remove Base's Foundation.
Foundation was only introduced as a child of Base, to render the body of
the Base, so that this could appear in front of the turret. Parents like
Base appear behind their children.

However, TIL there is a `.show_behind_parent' flag on Nodes. Setting
this on Turret means that Foundation is no longer required. A Base can
simply draw itself.