Fix collisions with shot and missile.
Previously I modified the radius of the bangshot's collision
circle to match the drawn expanding explosion. But Godot's
physics engine doesn't like modifying collision shapes on the
fly like this, and the collisions were inaccurate. Sometimes
missiles would fly through explosions visibly, or sometimes
be detonated by an explosion they didn't touch.

Now the collision shape is a fixed size - the max radius of
the explosion. When a missile enters this collision radius,
instead of destroying it immediately, the bangshot adds the
missile to a collection. The bangshot's _process tests each
missile in this collection to check whether it is closer than
the current drawn radius.

Along the way, messing with collision shapes and their layers,
masks, and signal connections, I've broken the collision between
missiles and the ground. But this will be easier to fix than the
above.