Fix bangshot's removal of freed missiles
Each bangshot maintains a collection of 'nearby_missiles'. Although a
bangshot correctly removes a missile from its collection when the
missile leaves the area, it does not account for a missile that is freed
by being destroyed (by collision with either another bangshot, or  the
ground.)

Such a freed missile remains in the collections of any bangshots it was
nearby to when it was freed. I had previously attempted to detect this
in _process, which iterates over the collection, with:

    if not is_instance_valid(misssile):

which worked, and then remove the missile from self..nearby_missiles
collection, with:

    self.nearby_missiles.erase(missile)

which did not. Apparently arrays can't erase freed objects. It gave an
error in the debugger output instead.

So now we instead copy valid (i.e. unfreed) missiles into a new
collection as we iterate over them, and swap that list of
'valid_missiles' into 'nearby_missiles' when we're done.