Replace object is null check with is_instance_valid
This is a bit of a stab in the dark. The line *before* has once
issued a warning: A missile executing this had no trail.
And is_instance_valid is the correct way to check for objects
which might be null or might have been marked for deletion.
But since this change is in the line *after* the one that issued
a non-fatal warning, I don't think I'm really doing anything
constructive here.