F3 now ends the level even if all bombs have not been launched yet
This has a wrinkle, in that if it is pressed at the start of the
level, before any bombs are created, then the level's "bombs remaining"
counter is set to zero, and there are no bombs in the air to destroy,
so the "last bomb destroyed" signal never gets emitted. Therefore
pressing F3 also emits this signal, ending the level.

However if it's pressed too soon, then the listener for this signal
hasn't started yet, so you might need to press F3 again, later,
for it to take effect.

Gah! Reality is a mess.